+++ bepasty/app.py
@@ -2,6 +2,7 @@
 # License: BSD 2-clause, see LICENSE for details.
 
 import os
+import time
 
 from flask import Flask, render_template
 
@@ -10,6 +11,7 @@ from .storage import create_storage
 from .views import blueprint
 from .rest_api import rest_api
 from .utils.name import setup_werkzeug_routing
+from .utils.permissions import *
 
 
 def create_app():
@@ -25,32 +27,33 @@ def create_app():
     app.register_blueprint(blueprint)
     app.register_blueprint(rest_api)
 
+    @app.errorhandler(403)
+    def page_not_found(e):
+        return render_template('_error_403.html'), 403
+
     @app.errorhandler(404)
     def page_not_found(e):
         return render_template('_error_404.html'), 404
 
+    def datetime_format(ts):
+        """
+        takes a unix timestamp and outputs a iso8601-like formatted string.
+        times are always UTC, but we don't include the TZ here for brevity.
+        it should be made clear (e.g. in the template) that the date/time is UTC.
+        """
+        if not ts:  # we use 0 to indicate undefined time
+            return 'undefined'
+        return time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime(ts))
+
+    app.jinja_env.filters['datetime'] = datetime_format
+
+    app.jinja_env.globals['logged_in'] = logged_in
+    app.jinja_env.globals['get_permissions'] = get_permissions
+    app.jinja_env.globals['may'] = may
+    app.jinja_env.globals['PERMISSIONS'] = PERMISSIONS
+    app.jinja_env.globals['ADMIN'] = ADMIN
+    app.jinja_env.globals['CREATE'] = CREATE
+    app.jinja_env.globals['READ'] = READ
+    app.jinja_env.globals['DELETE'] = DELETE
+
     return app
-
-
-def server_cli():
-    """Create command-line interface for bepasty server"""
-
-    import argparse
-
-    parser = argparse.ArgumentParser(
-        description="The free and open-source paste bin and trash can "
-                    "for your stuff.")
-    parser.add_argument('--host', help='Host to listen on')
-    parser.add_argument('--port', type=int, help='Port to listen on')
-    parser.add_argument('--debug', help='Activate debug mode',
-                        action='store_true')
-    args = parser.parse_args()
-
-    app = create_app()
-
-    print " * Starting bepasty server..."
-    app.run(
-        host=args.host,
-        port=args.port,
-        debug=args.debug
-    )