+++ bepasty/utils/hashing.py
@@ -1,0 +1,19 @@
+# Copyright: 2014 Thomas Waldmann <tw@waldmann-edv.de>
+# License: BSD 2-clause, see LICENSE for details.
+
+from hashlib import sha256 as hash_new
+
+SIZE = 1024 * 1024
+
+
+def compute_hash(data, size):
+    """
+    compute hash of storage item's data file, return hexdigest
+    """
+    hasher = hash_new()
+    offset = 0
+    while offset < size:
+        buf = data.read(SIZE, offset)
+        offset += len(buf)
+        hasher.update(buf)
+    return hasher.hexdigest()