+++ bepasty/views/display.py
@@ -2,8 +2,9 @@
 # License: BSD 2-clause, see LICENSE for details.
 
 import errno
+import time
 
-from flask import current_app, render_template, Markup, request, url_for
+from flask import current_app, render_template, Markup, request, url_for, abort
 from flask.views import MethodView
 from werkzeug.exceptions import NotFound
 from pygments import highlight
@@ -12,30 +13,50 @@ from pygments.formatters import HtmlFormatter
 
 
 from ..utils.name import ItemName
+from ..utils.permissions import *
 from . import blueprint
+from .filelist import file_infos
 
 
 class DisplayView(MethodView):
     def get(self, name):
+        if not may(READ):
+            abort(403)
         try:
-            item = current_app.storage.open(name)
+            item = current_app.storage.openwrite(name)
         except (OSError, IOError) as e:
             if e.errno == errno.ENOENT:
-                return render_template('file_not_found.html'), 404
+                abort(404)
+            raise
 
         with item as item:
-            if not item.meta.get('unlocked'):
-                error = 'File locked.'
-            elif not item.meta.get('complete'):
+            if not item.meta['complete']:
                 error = 'Upload incomplete. Try again later.'
             else:
                 error = None
             if error:
-                return render_template('display_error.html', name=name, item=item, error=error), 409
+                return render_template('error.html', heading=item.meta['filename'], body=error), 409
+
+            if item.meta['locked'] and not may(ADMIN):
+                abort(403)
+
+            def read_data(item):
+                # reading the item for rendering is registered like a download
+                data = item.data.read(item.data.size, 0)
+                item.meta['timestamp-download'] = int(time.time())
+                return data
+
             ct = item.meta['type']
-            if ct.startswith('text/'):
-                code = item.data.read(item.data.size, 0)
-                code = code.decode('utf-8')  # TODO we don't have the coding in metadata
+            if ct.startswith('text/x-bepasty-'):
+                # special bepasty items
+                if ct == 'text/x-bepasty-list':
+                    names = read_data(item).splitlines()
+                    files = sorted(file_infos(names), key=lambda f: f['filename'])
+                    rendered_content = Markup(render_template('filelist_tableonly.html', files=files))
+                else:
+                    rendered_content = u"Can't render this content type."
+            elif ct.startswith('text/'):
+                code = read_data(item).decode('utf-8')  # TODO we don't have the coding in metadata
                 lexer = get_lexer_for_mimetype(ct)
                 formatter = HtmlFormatter(linenos='table', lineanchors="L", anchorlinenos=True)
                 rendered_content = Markup(highlight(code, lexer, formatter))